0.1.8

+++ Source/Delete.ts
@@ -35,25 +35,31 @@ export default async (
 	const Deleted = [];
 
 	for (const { name } of (await Project(ID, Header)) ?? []) {
-		for (const { id } of (
+		for (const { id, created_on } of (
 			await (async (Project: string) =>
 				(await Deployment(ID, Project, Header)).splice(0, Limit) ?? [])(
 				name
 			)
 		).reverse()) {
-			try {
-				await fetch(
-					`${`https://api.cloudflare.com/client/v4/accounts/${ID}/pages/projects/${name}/deployments`}/${id}`,
-					{
-						method: "DELETE",
-						headers: Header,
-					}
-				);
-			} catch (_Error) {
-				console.log(_Error);
-			}
+			if (
+				// @ts-ignore
+				(Date.now() - new Date(created_on)) / 86400000 >
+				Days
+			) {
+				try {
+					await fetch(
+						`${`https://api.cloudflare.com/client/v4/accounts/${ID}/pages/projects/${name}/deployments`}/${id}`,
+						{
+							method: "DELETE",
+							headers: Header,
+						}
+					);
+				} catch (_Error) {
+					console.log(_Error);
+				}
 
-			Deleted.push(id);
+				Deleted.push(id);
+			}
 		}
 	}
